chore: refactoring, bump zod
