+++ packages/generator-networks-creator/package.json
@@ -21,7 +21,7 @@
         "node-fetch": "^2.6.7",
         "tslib": "^2.4.0",
         "ua-parser-js": "^2.0.0",
-        "zod": "^3.24.4"
+        "zod": "^3.25.1"
     },
     "repository": {
         "type": "git",
